Remove unnecessary stopwatch in mouse gesture handler. Code cleanup for camera controller. Use ring buffer for camera history. Change camera history class to struct.

+++ Source/HelixToolkit.Wpf.SharpDX/Controls/MouseHandlers/PanHandler.cs
@@ -147,7 +147,7 @@ namespace HelixToolkit.Wpf.SharpDX
         /// <param name="elapsedTime">
         /// The elapsed time (milliseconds).
         /// </param>
-        protected override void OnInertiaStarting(int elapsedTime)
+        protected override void OnInertiaStarting(double elapsedTime)
         {
             var speed = (this.LastPoint - this.MouseDownPoint) * (40.0 / elapsedTime);
             this.Viewport.AddPanForce(speed.X, speed.Y);
